Add a parser for Green trees
This commit adds a basic parser which takes a list of tokens and turns them into a
structured green tree. The parser currently supports properly formatted S-expressions
(i.e. the S-expressions must have balanced parens) and creates a struture in the
green tree that matches the nesting in the input.