Add iterator for traversal of red node children
This commit adds a children method to RedNodeData which allows for the traversal
of the nodes in both directions. The RedNodes are built on the fly when
traversing the tree top down, the red nodes wrap green nodes but also contain a
reverence to their parent. This allows you to walk the tree downward using the
children contained in green nodes and then follow the parent pointers contained
in red nodes back up to the root.