Add another test that checks tree traversal of children
This commit adds another test that checks what happens when traversing children
that contain both nodes and tokens. Calling .children() on Tokens returns an empty
iterator, calling it on nodes will return a RedNode for each token/node at that
level and the red node will contain a reference to its parent.