Have NodeData.children() return iterator
This commit changes the NodeData.children method so that it returns an iterator
over the children of a node. This is useful because it allows the caller to then
use iterator methods to access the children of a green node. Prior to this commit
the caller would get the slice returned from NodeData.children and would then
turn that into an iterator with an additional .iter() call.