Add replace_child method to RedNodeData
This commit adds a replace_child method for RedNodes which takes a child index
and a new green node and will return a new root RedNode where the underlying
green tree has been updated. When navigating a red tree this method can be
called at any RedNode, but will return None if called on a RedToken because
RedTokens don't have children.