Add Note about children access in RedNodeData
This commit adds a note about the access to child nodes in this naive impl
of syntax trees. In the real Rowen the offset_in_parent is stored per child
in the green node's children vector. This difference allows for the efficient
access of children when using pointers to recreate syntax trees.