Add separate enum to wrap GreenNode/Token referenced in RedNode
This commit adds a separate enum to wrap GreenNodes/Tokens which are referenced
in RedNodes. This wrapping is needed to allow RedNodes to handle both types
found in green trees and helps avoid a bunch of duplication in the structs for
red trees. Prior to this commit there were cases when you would have to unwrap
the NodeOrToken enum twice when processing a RedNode, once for the GreenNode
referenced by the RedNode and then again for any children found in the GreenNode.