Allow green tree parser to emit errors and create partial green nodes
This commit allows the basic green node parser to emit errors when an unmatched
right paren is found. The error doesn't cause the parser to hard fail. A partial
green node is produced instead and the error is marked. This allows the parser
to still create a green tree that can reproduce the original text.