Add basic AST parser for RedNodes
This commit adds a basic parser that can take a RedNode containing an
sexpression and wraps the RedNode in a Ast::Expr type, extracting information
about the node in the process. This allows you to have a strictly typed tree
while still maintaining access to the dynamically typed Rowan syntax tree
that the AST is built on top of.